Moved chandelier indicator to specific package
